==> 1.Ambient_RNA_Removal.R <==
#PART 1: AMBIENT RNA REMOVAL

library(dplyr)
library(Seurat)
library(patchwork)
library(SoupX)
library(stringr)

#Dataset Pathway Setup
setwd("~/20_015")

# Cell-Cycle Scoring & Regression Gene List
s.genes <- cc.genes.updated.2019$s.genes
g2m.genes <- cc.genes.updated.2019$g2m.genes

#Change Human Gene to Mouse Gene Nomenclature
g2m.genes <- str_to_title(g2m.genes) 
s.genes <- str_to_title(s.genes) 

#Removal of Doublets & Ambient RNA
##DAY0_UNL
tod <- Read10X(data.dir = "Day0_UNL/raw_feature_bc_matrix/")
toc <- Read10X(data.dir = "Day0_UNL/filtered_feature_bc_matrix/")
sc = SoupChannel(tod, toc)
srat = CreateSeuratObject(sc$toc)
srat <- NormalizeData(srat)
srat <- FindVariableFeatures(srat, nfeatures = 4000)
srat <- CellCycleScoring(srat, s.features = s.genes, g2m.features = g2m.genes, set.ident = TRUE)
srat <- ScaleData(srat, vars.to.regress = c("S.Score", "G2M.Score"), verbose = TRUE)
srat <- RunPCA(srat, npcs = 50, verbose = TRUE)
srat <- FindNeighbors(srat, reduction = "pca", dims = 1:50)
srat <- FindClusters(srat, resolution = 1)
sc <- setClusters(sc,srat@active.ident)
sc <- autoEstCont(sc)
DAY0_UNL.deSouped <- adjustCounts(sc)
DAY0_UNL <- CreateSeuratObject(DAY0_UNL.deSouped, project = "DAY0_UNL")
saveRDS(DAY0_UNL, file = "souped_DAY0_UNL.rds")

##DAY0
tod <- Read10X_h5("Day0/raw_feature_bc_matrix.h5")
toc <- Read10X_h5("Day0/filtered_feature_bc_matrix.h5")
sc = SoupChannel(tod, toc)
srat = CreateSeuratObject(sc$toc)
srat <- NormalizeData(srat)
srat <- FindVariableFeatures(srat, nfeatures = 4000)
srat <- CellCycleScoring(srat, s.features = s.genes, g2m.features = g2m.genes, set.ident = TRUE)
srat <- ScaleData(srat, vars.to.regress = c("S.Score", "G2M.Score"), verbose = TRUE)
srat <- RunPCA(srat, npcs = 50, verbose = TRUE)
srat <- FindNeighbors(srat, reduction = "pca", dims = 1:50)
srat <- FindClusters(srat, resolution = 1)
sc <- setClusters(sc,srat@active.ident)
sc <- autoEstCont(sc)
DAY0.deSouped <- adjustCounts(sc)
DAY0 <- CreateSeuratObject(DAY0.deSouped, project = "DAY0")
saveRDS(DAY0, file = "souped_DAY0.rds")

##DAY0.5
tod <- Read10X(data.dir = "Day0.5/raw_feature_bc_matrix/")
toc <- Read10X(data.dir = "Day0.5/filtered_feature_bc_matrix/")
sc = SoupChannel(tod, toc)
srat = CreateSeuratObject(sc$toc)
srat <- NormalizeData(srat)
srat <- FindVariableFeatures(srat, nfeatures = 4000)
srat <- CellCycleScoring(srat, s.features = s.genes, g2m.features = g2m.genes, set.ident = TRUE)
srat <- ScaleData(srat, vars.to.regress = c("S.Score", "G2M.Score"), verbose = TRUE)
srat <- RunPCA(srat, npcs = 50, verbose = TRUE)
srat <- FindNeighbors(srat, reduction = "pca", dims = 1:50)
srat <- FindClusters(srat, resolution = 1)
sc <- setClusters(sc,srat@active.ident)
sc <- autoEstCont(sc)
DAY0.5.deSouped <- adjustCounts(sc)
DAY0.5 <- CreateSeuratObject(DAY0.5.deSouped, project = "DAY0.5")
saveRDS(DAY0.5, file = "souped_DAY0.5.rds")

##DAY1_LIG
tod <- Read10X(data.dir = "Day1_LIG/raw_feature_bc_matrix/")
toc <- Read10X(data.dir = "Day1_LIG/filtered_feature_bc_matrix/")
sc = SoupChannel(tod, toc)
srat = CreateSeuratObject(sc$toc)
srat <- NormalizeData(srat)
srat <- FindVariableFeatures(srat, nfeatures = 4000)
srat <- CellCycleScoring(srat, s.features = s.genes, g2m.features = g2m.genes, set.ident = TRUE)
srat <- ScaleData(srat, vars.to.regress = c("S.Score", "G2M.Score"), verbose = TRUE)
srat <- RunPCA(srat, npcs = 50, verbose = TRUE)
srat <- FindNeighbors(srat, reduction = "pca", dims = 1:50)
srat <- FindClusters(srat, resolution = 1)
sc <- setClusters(sc,srat@active.ident)
sc <- autoEstCont(sc)
DAY1_LIG.deSouped <- adjustCounts(sc)
DAY1_LIG <- CreateSeuratObject(DAY1_LIG.deSouped, project = "DAY1_LIG")
saveRDS(DAY1_LIG, file = "souped_DAY1_LIG.rds")

##DAY1
todL <- Read10X(data.dir = "Day1/raw_feature_bc_matrix/")
toc <- Read10X(data.dir = "Day1/filtered_feature_bc_matrix/")
sc = SoupChannel(tod, toc)
srat = CreateSeuratObject(sc$toc)
srat <- NormalizeData(srat)
srat <- FindVariableFeatures(srat, nfeatures = 4000)
srat <- CellCycleScoring(srat, s.features = s.genes, g2m.features = g2m.genes, set.ident = TRUE)
srat <- ScaleData(srat, vars.to.regress = c("S.Score", "G2M.Score"), verbose = TRUE)
srat <- RunPCA(srat, npcs = 50, verbose = TRUE)
srat <- FindNeighbors(srat, reduction = "pca", dims = 1:50)
srat <- FindClusters(srat, resolution = 1)
sc <- setClusters(sc,srat@active.ident)
sc <- autoEstCont(sc)
DAY1.deSouped <- adjustCounts(sc)
DAY1 <- CreateSeuratObject(DAY1.deSouped, project = "DAY1")
saveRDS(DAY1, file = "souped_DAY1.rds")


##DAY4
tod <- Read10X(data.dir = "Day4/raw_feature_bc_matrix/")
toc <- Read10X(data.dir = "Day4/filtered_feature_bc_matrix/")
sc = SoupChannel(tod, toc)
srat = CreateSeuratObject(sc$toc)
srat <- NormalizeData(srat)
srat <- FindVariableFeatures(srat, nfeatures = 4000)
srat <- CellCycleScoring(srat, s.features = s.genes, g2m.features = g2m.genes, set.ident = TRUE)
srat <- ScaleData(srat, vars.to.regress = c("S.Score", "G2M.Score"), verbose = TRUE)
srat <- RunPCA(srat, npcs = 50, verbose = TRUE)
srat <- FindNeighbors(srat, reduction = "pca", dims = 1:50)
srat <- FindClusters(srat, resolution = 1)
sc <- setClusters(sc,srat@active.ident)
sc <- autoEstCont(sc)
DAY4.deSouped <- adjustCounts(sc)
DAY4 <- CreateSeuratObject(DAY4.deSouped, project = "DAY4")
saveRDS(DAY4, file = "souped_DAY4.rds")


##DAY7
tod <- Read10X(data.dir = "Day7/raw_feature_bc_matrix/")
toc <- Read10X(data.dir = "Day7/filtered_feature_bc_matrix/")
sc = SoupChannel(tod, toc)
srat = CreateSeuratObject(sc$toc)
srat <- NormalizeData(srat)
srat <- FindVariableFeatures(srat, nfeatures = 4000)
srat <- CellCycleScoring(srat, s.features = s.genes, g2m.features = g2m.genes, set.ident = TRUE)
srat <- ScaleData(srat, vars.to.regress = c("S.Score", "G2M.Score"), verbose = TRUE)
srat <- RunPCA(srat, npcs = 50, verbose = TRUE)
srat <- FindNeighbors(srat, reduction = "pca", dims = 1:50)
srat <- FindClusters(srat, resolution = 1)
sc <- setClusters(sc,srat@active.ident)
sc <- autoEstCont(sc)
DAY7.deSouped <- adjustCounts(sc)
DAY7 <- CreateSeuratObject(DAY7.deSouped, project = "DAY7")
saveRDS(DAY7, file = "souped_DAY7.rds")
